Refactor hardware management out of bootstrap.rs
